msm8996-common: ConfigPanel: Refactor
 * Integrate more nicely with Settings.
 * Remove unnecessary permissions and unused resources.
 * Use cmsdk FileUtils in all cases where it suits and ensure that
   files are writable before actually writing to them.

+++ configpanel/src/com/cyanogenmod/settings/device/ButtonSettingsFragment.java
@@ -15,52 +15,46 @@
  * limitations under the License.
  */
 
-package com.cyanogenmod.settings.device.utils;
+package com.cyanogenmod.settings.device;
 
 import android.os.Bundle;
-import android.preference.Preference;
-import android.preference.Preference.OnPreferenceChangeListener;
-import android.preference.PreferenceActivity;
-import android.preference.ListPreference;
-import android.preference.SwitchPreference;
+import android.support.v14.preference.PreferenceFragment;
+import android.support.v14.preference.SwitchPreference;
+import android.support.v7.preference.ListPreference;
+import android.support.v7.preference.Preference;
+import android.support.v7.preference.Preference.OnPreferenceChangeListener;
 import android.text.TextUtils;
 import android.view.MenuItem;
 
-import java.io.File;
+import com.android.settingslib.drawer.SettingsDrawerActivity;
+import com.cyanogenmod.settings.device.utils.Constants;
 
 import org.cyanogenmod.internal.util.FileUtils;
-import org.cyanogenmod.internal.util.ScreenType;
 
-public class NodePreferenceActivity extends PreferenceActivity
+public class ButtonSettingsFragment extends PreferenceFragment
         implements OnPreferenceChangeListener {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        getActionBar().setDisplayHomeAsUpEnabled(true);
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        addPreferencesFromResource(R.xml.button_panel);
     }
 
     @Override
-    protected void onResume() {
+    public void onResume() {
         super.onResume();
         updatePreferencesBasedOnDependencies();
-
-        // If running on a phone, remove padding around the listview
-        if (!ScreenType.isTablet(this)) {
-            getListView().setPadding(0, 0, 0, 0);
-        }
     }
 
     @Override
     public boolean onPreferenceChange(Preference preference, Object newValue) {
         String node = Constants.sBooleanNodePreferenceMap.get(preference.getKey());
-        if (!TextUtils.isEmpty(node)) {
+        if (!TextUtils.isEmpty(node) && FileUtils.isFileWritable(node)) {
             Boolean value = (Boolean) newValue;
             FileUtils.writeLine(node, value ? "1" : "0");
             return true;
         }
         node = Constants.sStringNodePreferenceMap.get(preference.getKey());
-        if (!TextUtils.isEmpty(node)) {
+        if (!TextUtils.isEmpty(node) && FileUtils.isFileWritable(node)) {
             FileUtils.writeLine(node, (String) newValue);
             return true;
         }
@@ -76,7 +70,7 @@ public class NodePreferenceActivity extends PreferenceActivity
             if (b == null) continue;
             b.setOnPreferenceChangeListener(this);
             String node = Constants.sBooleanNodePreferenceMap.get(pref);
-            if (new File(node).exists()) {
+            if (FileUtils.isFileReadable(node)) {
                 String curNodeValue = FileUtils.readOneLine(node);
                 b.setChecked(curNodeValue.equals("1"));
             } else {
@@ -88,7 +82,7 @@ public class NodePreferenceActivity extends PreferenceActivity
             if (l == null) continue;
             l.setOnPreferenceChangeListener(this);
             String node = Constants.sStringNodePreferenceMap.get(pref);
-            if (new File(node).exists()) {
+            if (FileUtils.isFileReadable(node)) {
                 l.setValue(FileUtils.readOneLine(node));
             } else {
                 l.setEnabled(false);
@@ -96,27 +90,16 @@ public class NodePreferenceActivity extends PreferenceActivity
         }
     }
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-        // Respond to the action bar's Up/Home button
-        case android.R.id.home:
-            finish();
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
     private void updatePreferencesBasedOnDependencies() {
         for (String pref : Constants.sNodeDependencyMap.keySet()) {
             SwitchPreference b = (SwitchPreference) findPreference(pref);
             if (b == null) continue;
             String dependencyNode = Constants.sNodeDependencyMap.get(pref)[0];
-            if (new File(dependencyNode).exists()) {
+            if (FileUtils.isFileReadable(dependencyNode)) {
                 String dependencyNodeValue = FileUtils.readOneLine(dependencyNode);
                 boolean shouldSetEnabled = dependencyNodeValue.equals(
                         Constants.sNodeDependencyMap.get(pref)[1]);
-                Constants.updateDependentPreference(this, b, pref, shouldSetEnabled);
+                Constants.updateDependentPreference(getContext(), b, pref, shouldSetEnabled);
             }
         }
     }
